Rename retained IO preamble boundary
