+++ crates/sifr_codegen/src/preamble.rs
@@ -28,7 +28,7 @@ mod process_child_pipes;
 mod process_runtime;
 pub(crate) use process_async_runtime::*;
 pub(crate) use process_runtime::*;
-mod io_logging_random;
-pub use io_logging_random::*;
+mod io_file_handles;
+pub use io_file_handles::*;
 mod io_bytes_methods;
 pub(crate) use io_bytes_methods::*;
