+++ plans/reviews/active/m4d-preamble-cleanup-opus-round1.md
@@ -1,0 +1,25 @@
+# M4d Preamble Cleanup Review - Opus Round 1
+
+Status: READY
+
+Reviewer command:
+
+```bash
+claude --print --dangerously-skip-permissions --setting-sources project --model claude-opus-4-7 --effort xhigh "<bounded diff review prompt>"
+```
+
+Findings:
+
+- No blockers found.
+- The remaining generated preamble file is IO/file-handle only and has been renamed from `io_logging_random.rs` to `io_file_handles.rs`.
+- `preamble.rs` module wiring matches the rename.
+- The retained manifest keeps the historical row ID for schema continuity while updating `preamble_files` and the reason to record that logging/random migrated out.
+- Remaining `io_logging_random` mentions are historical review artifacts and were intentionally left untouched.
+
+Validation evidence provided to reviewer:
+
+- `cargo fmt --check`
+- `python3 scripts/check_stdlib_manifest_schema.py`
+- `python3 scripts/check_stdlib_native_intrinsic_allowlist.py`
+- `python3 scripts/check_stdlib_migration_closure.py`
+- `cargo test -p sifr_codegen io_bytes_methods`
